Increases gen_server:call timeout to :infinity
Increases timeout from 5000ms to :infinity. In
environments with a low pool size, e.g. CI workers
and lesser powered laptops, test failures due to a
delay > 5000ms in checking out a worker from the
pool frequently occur. As ExUnit has a default max
time for a test case to complete (60_000ms), it
makes sense to avoid these failures by allowing
a test to block until a worker is checked back in
to the pool.