Fix dialyzer on travis ci
This runs the build with sudo: required so it runs on a larger vm with
7.5GB of memory, so mix dialyzer --plt doesn't run the vm out of memory
and exit with code 137.
See:
https://docs.travis-ci.com/user/ci-environment/#Virtualization-environments

This also runs wraps the mix dialyzer --plt in the travis_wait command
so if it doesn't output any log statements in 10 minutes, the build
won't timeout.

Finally, the build is already being cached, so the only time the PLT
needs to be built is on a fresh cache, or if the build matrix is
changed.

Fixes #186.