Make Phantom.Server escript compatible (#193)
This allows Wallaby.Phantom.Server to run when priv/run_phantom.sh isn't
available (like when running as an escript). Since we can't directly run
the script, we need to run /bin/sh with the arguments we want to use and
pipe the contents of run_phantom.sh in via stdin. This is almost the
same script as in the Elixir Port docs
(https://hexdocs.pm/elixir/Port.html#module-zombie-processes), but it
runs the command in the background and then listens for stdin to close
and kills the process from a subshell.

I have confirmed this shuts down phantom in the following situations:

* ^C+^C in IEx
* causing the beam to core dump by overflowing the atom table
* running kill -9 on the beam.

Huge thanks to @koudelka for all of his help coming up with this
approach.

Fixes #192