Move current driver into Wallaby.Session
This moves the responsibility for determining the Wallaby driver to use
to the Session and breaks all of the hard-coded references to
Wallaby.Phantom.Driver.

I also created a Wallaby.Driver behaviour which new drivers will need to
implement. I left this behaviour `@moduledoc false` in case we wanted to
adjust the interface before making it public.