Fix leaking webdriver sessions in QueryTest
Wallaby.Integration.QueryTest was starting a ton of sessions and not
cleaning them up. This switches it over to use
Wallaby.Integration.SessionCase and uses the session that's
automatically created and cleaned up. I also extracted the unit tests
out of this integration test so they can be run separately.