Experimental selenium driver (#198)
* Experimental Selenium driver

Updates to build

* Handle 204 response from webdriver api

I have to return %{“value” => nil} because everywhere uses 
Map.fetch(body, “value”) and if I didn’t have a value key in the map, 
this would errors. This API really needs to be changed.

* Fix integration tests for selenium

I added page objects so I have somewhere to add checking to see if the
page is loaded yet. With selenium, when we send a visit command, it
doesn't wait until the page is completely loaded to run the next
command.

* Fix test failure on a small screen

If the screen height was less then 1234 pixels, the test would fail.
This uses a much smaller screen height.

* Fixes issues with click_button test

There were a couple issues with click_button test.

1. Selenium wasn't waiting until the page finished loading to get the
   current url, so these tests were failing
2. When the form was submitted it sent a POST request to the test
   server, which resulted in a 500 error. I updated it to send a get
   request, but then had to update the current url assertion since the
   form data was being included as url parameters.

* Fixing integration test on Selenium 2

The older version of Firefox injects the xmlns attribute in the html
tag. We could probably improve this test if it continues to be an issue.

* Renamed PageObjects to Pages

Also fixed an instance where I was using a deprecated api

* Set default Session.server to :none

See https://github.com/keathley/wallaby/pull/198#discussion_r113946313

* Retry if can't start test session

I have been having issues on travis ci where selenium times out when
starting a session. This should take care of it.