Fixes visit/2 issue where urls have too many slashes
I noticed in #300 that visit/2 is generating urls like
http://localhost:54579//logs.html that have too many slashes. This
occurs when the base_url has a trailing slash and the requested path
has a leading slash. This change adds tests and fixes the issue so we
don't end up with urls with too many or not enough slashes
depending on how these settings are configured.