Capture JS logs in Chrome (#300)
Issue #278

This commit captures js logs for chromedriver in the same way that logs
were captured for phantom.  That is, after each action the driver checks
for logs, appends to the LogStore and parses the logs to be output to
the terminal.

The LogStore and LogChecker have been abstracted for use by all drivers.
Each driver implements its own Logger module.

The chrome driver was delegating to the selenium driver
in many cases, making it hard to wrap the function with a `check_logs!`
call in the same way that the phantom driver did it.  To accomodate
that, we created a private `delegate` function that calls `check_logs`
and the calls `Kernel.apply` to call the appropriate driver function.