Upgrade Selenium Server and GeckoDriver for CI (#392)
* Update selenium server and geckodriver for WALLABY_SELENIUM_VERSION=3

* Use new web element identifier when pattern matching on element

According to the [WebDriver
specification](https://www.w3.org/TR/webdriver/#elements), elements will
be returned with the web element identifier as a property on the
response object. This is a constant of
`element-6066-11e4-a52e-4f735466cecf`. I am not sure of the significance
of this constant, however.