Add selected condition to query
Description
===========

Adds a `selected` condition to queries. This should only work with
options, radio buttons, and checkboxes (since it uses
`Element.selected?/1` underneath).

The usage is similar to the `visible` option for a query function: e.g.
`Query.css("selector", selected: true)`. The default case, when neither
`selected: true` nor `selected: false` is passed sets the condition to
`:any`. In this way, if neither `true` nor `false` are specified, the
query will return both selected and unselected elements.

The error message was also updated to include the notion of whether
selected or unselected elements were being searched for.