Make has_text? and assert_text typespecs more general (#559)
Dialyzer throws warnings when you call `has_text?` and `assert_text` with a `session`, although it is supported. Update `@spec` to ensure dialyzer knows what's up.

The `locator` type def is only used 3 times, and is just an alias for `String.t()`, which is used 33 times in this module. Drop `locator` in preference for sticking with `String.t()`.

Co-authored-by: Michał Łępicki <michallepicki@users.noreply.github.com>