Ensure all sessions are closed on exit
If you were running a test with many sessions, and it was either the
last test to run or the only test to run, then there was the chance
that the SessionStore would exit before we could ask the webdriver
to close the sessions.

While the webdriver would successfully die, this would leave zombie
browser processes lingering, tormenting the living.

Here, we use the after_suite callback from ExUnit to ensure that these
sessions get closed before ExUnit kills Wallaby.

Unfortunately this was introduced in Elixir 1.8, so we now have to raise
the minimum Elixir version to 1.8, which I think is reasonable, considering
the latest version is 1.11.