Wallaby.Element: handle errors more consistently (#633)
Fixes #631

A number of functions in Wallaby.Element did not handle {:error, _}
tuples, which meant that they’re raise a CaseClauseError if they
received an error from the driver.

This commit extracts some private functions to make the handling of
driver results more concise and consistent.

I didn’t see any point adding more tests for StaleReferenceError, as
there is already a test for this error, and everything now uses the same
code path.

I removed InvalidSelectorError -- this was only used by
Wallaby.Element.clear/1. As Wallaby.Element is dealing
with elements that we have already found, it’s not possible that we’ll
be handling an invalid selector at this point.

Wallaby.Browser does deal with invalid selectors, but they are raised as
a QueryError, rather than InvalidSelectorError. (There were also no
tests for the latter.)