add check for chromedriver chrome version match (#688)
* Fixes #685
When google chrome and chromedriver versions are out of sync wallaby can start in a bad state.
We added a version comparison step to `Wallaby.Chrome.validate/0` that will error
during application start if chrome and chrome versions do not match to a major, minor, and build version.

Co-authored-by: Alex McLain <alex@alexmclain.com>
Co-authored-by: Johnny Otsuka <johnny@binarynoggin.com>
Co-authored-by: John Cotton <cotton@binarynoggin.com>
Co-authored-by: Liam Elder <liam@binarynoggin.com>

* Use correct config option in DependencyError

To configure the Chrome binary, you use the `:chromdriver` option and
set it to `[binary: "/path/to/chrome"]`

* Emit a warning instead of raising an error

Sometimes having a mismatched chromedriver/chrome version still works,
and having this error might inconvenience the developer. Emitting a
warning I think is sufficient to alert the developer as to a potential
problem.

This also fixes some incorrect config value. The path to an alternative
chrome binary is found is found with `Application.get_env(:wallaby,
:chromedriver)[:binary]`, not
`Application.get_env(:wallaby, :chrome)[:path]`.

Co-authored-by: Matt Hall <401matthall@gmail.com>
Co-authored-by: Alex McLain <alex@alexmclain.com>
Co-authored-by: Johnny Otsuka <johnny@binarynoggin.com>
Co-authored-by: John Cotton <cotton@binarynoggin.com>
Co-authored-by: Liam Elder <liam@binarynoggin.com>