Start chromedriver with PartitionSupervisor
This will open as many chromedrivers you have as cores, and partition
them access to them based on your test pid. This effectively acts as a
"pool" of chromedrivers without needing logic to check them in and out
of the pool.

It just acts as if there is a single chromedriver and it seems to just
work.

Related to #365