Add sass asset note to README
The **Assets/esbuild** section of the README includes a step on adding `esbuild default` to the `test` alias. This works to recompile javascript files on every run; but it only compiles scss styles on the *first* run, leading to continued confusion when debugging.

To do this, we simply need to add the `sass default` step to the `test` alias.