Start chromedriver with PartitionSupervisor (#692)
* Start chromedriver with PartitionSupervisor

This will open as many chromedrivers you have as cores, and partition
them access to them based on your test pid. This effectively acts as a
"pool" of chromedrivers without needing logic to check them in and out
of the pool.

It just acts as if there is a single chromedriver and it seems to just
work.

Related to #365

* ci: increase minimum elixir/otp versions

* fix: add license for PartitionSupervisor

Since we are vendoring this from the Elixir codebase, we attribute the
project by providing its license in the file.

* chore: format

Formatting has changed since we have updated the minimum Elixir version.

* chore: Add link to elixir-lang/elixir repo for vendored code

* deps: Update Credo

* chore: add NOTICE.md and update LICENSE.md

* fix: Version mismatch of how to call IO.binread

It seems that using `:all` with IO.binread has changed between v1.12 and
v1.13, and is causing a Dialyzer problem. This patch adds some compile
time logic to call it with the right argument based on the Elixir
version.