Fixes misreported HTML errors when input labels are changed by CSS
This fixes #236 by using the validate_html logic, that was previously
used exclusively by buttons, for both checkboxes and radio buttons.
This changes the error message Wallaby reports from complaining about
a missing label to one that simply states the label does not exist.

This also extends the HTML validation for buttons so that helpful hints
can still be provided, and adds a new error message for when the element
isn't found but the CSS query finds the label.