Add partition count config and improved defaults
The user can explicitly set the number of chromedriver partitions now.

But even if the user does not set this config, Wallaby will make an
improved choice as to how many partitions it fires up.

This will lead to better performance and fewer false negative test
results on machines with high CPU-core counts.