set Wallaby test concurrency same as user value
Without this change, running mix test on Wallaby itself on a 56-core
machine produces the same undesirable side effects we are trying to
remedy with this PR.

Specifically, all the CPU cores fire up, tests complete slowly (at
least 45 seconds for a single test), and with many errors.

We believe this is due to something related to how chromedriver
operates, but the workaround of limiting concurrency to 10 processes
works well enough for us to defer going down that rabbit hole.