fix: fix wrong option and remove partition supervisor change
I had missed that the `:partitions` option was changed to `:concurrency`
when defing the partition supervisor. This was not doing anything, which
is why making the change in PartitionSupervisor was necessary.

Now that the supervisor is defined with the correct option, the change
in the (vendored) PartitionSupervisor module is not necessary.