Only spawn a browser when using the `feature` macro, not normal `test`s (#795)
* Only spawn a browser when using the `feature` macro, not normal `test`s

This fixes an unintended behavior where having `use Wallaby.Feature` at the top of your module would cause a browser to be spawned for every test in the module, not just `feature` tests. The result is that if you had a large test module, most of which used the normal `test` macro and one of which used the `feature` macro, every test would take a minimum of a third of a second or more.

I also took the liberty to check and see in the `setup` block if we already had a `:session`, and if so, avoid spawning a second. I don't *think* anyone was relying on the behavior where a second browser was being spawned if you had `use Wallaby.Feature` at both the top of the module and a `describe` block.

Resolves #794

* Bump upload-artifact version to fix deprecation error

* Address PR feedback